% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comboChunk.R
\name{comboChunk.ultraCombo}
\alias{comboChunk.ultraCombo}
\title{comboChunk.ultraCombo}
\usage{
\method{comboChunk}{ultraCombo}(combo)
}
\arguments{
\item{combo}{an 'ultraCombo' to chunk}
}
\description{
comboChunk.ultraCombo
}
