% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{integer.precision.limit}
\alias{integer.precision.limit}
\title{integer.precision.limit}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
integer.precision.limit
}
\description{
A constant. The maximum double value the base system can handle
with integer precision.\cr 
\code{
max(.Machine$integer.max,
2^.Machine$double.digits-1
)
/}
}
\keyword{datasets}
