% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOps.combo.R
\name{invert.combo}
\alias{invert.combo}
\title{invert.combo}
\usage{
invert.combo(a)
}
\arguments{
\item{a}{an ultraCombo object}
}
\description{
invert.combo
}
\examples{
n<-4
k<-2
a<-ultraCombo(vector(),n,k)
b<-invert.combo(a)
print(b)
print(b$i)
stopifnot(all(b$i==seq(6)))
}
