% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is.valid.index}
\alias{is.valid.index}
\title{is.valid.index}
\usage{
is.valid.index(i, n, k)
}
\arguments{
\item{i}{A vector of indices}

\item{n}{an integer, n of combination}

\item{k}{an integer, k of combination}
}
\description{
is.valid.index
}
\examples{
n<-20
k<-10
# non-integer input
try(is.valid.index(0.5,n,k))
# out of range input
try(is.valid.index(0,n,k))
try(is.valid.index(choose(n,k)+1,n,k))
# valid tests
# no argument
is.valid.index(vector(),n,k)
# single valid argument
is.valid.index(1,n,k)
# multiple valid aguments
is.valid.index(c(1,choose(n,k)),n,k)
}
