% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{print.ultraCombo}
\alias{print.ultraCombo}
\title{print.ultraCombo}
\usage{
\method{print}{ultraCombo}(x, ...)
}
\arguments{
\item{x}{An 'ultraCombo' to print}

\item{...}{ignored}
}
\description{
print.ultraCombo
}
