% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOps.combo.R
\name{setdiff.combo}
\alias{setdiff.combo}
\title{setdiff.combo}
\usage{
setdiff.combo(a, b)
}
\arguments{
\item{a, b}{one ultraCombo and one ultraCombo or vector}
}
\description{
setdiff.combo
}
\examples{
a<-ultraCombo(seq(4),4,2)
b<-ultraCombo(seq(3)+2,4,2)
a$i
b$i
stopifnot(all(c(1,2)==setdiff.combo(a,b)$i))
stopifnot(all(c(1,4)==setdiff.combo(a,seq(2)+1)$i))
stopifnot(all(c(1,2,6)==setdiff.combo(seq(6),b)$i))
}
