% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combo.R
\name{[.ultraCombo}
\alias{[.ultraCombo}
\title{[.ultraCombo}
\usage{
\method{[}{ultraCombo}(x, i)
}
\arguments{
\item{x}{a combo object}

\item{i}{index of the desired subset of combo object}
}
\description{
[.ultraCombo
}
\examples{
n<-4
k<-2
combo<-ultraCombo(seq(choose(n,k)),n,k)
combo[seq(3)]
combo[seq(3)*2]
try(combo[0])
try(combo[7])
}
