% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ultraCombo.R
\docType{package}
\name{ultraCombo-package}
\title{ultraCombo-package}
\description{
Functions and data storage and manipulation 
for combination objects.
}
\author{
\strong{Maintainer}: Dale \email{dale@piratepress.org} [copyright holder]

}
