% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOps.combo.R
\name{union.combo}
\alias{union.combo}
\title{union.combo}
\usage{
union.combo(...)
}
\arguments{
\item{...}{at least one ultraCombo object, and any amount of ultraCombos or vectors}
}
\description{
union.combo
}
\examples{
a<-ultraCombo(seq(4),4,2)
b<-ultraCombo(seq(3)+2,4,2)
a$i
b$i
stopifnot(all(seq(5)==union.combo(a,b)$i))
print(seq(4))
print(a$i)
print(seq(2)+1)
print(union.combo(a,seq(2)+1)$i)
stopifnot(all(seq(4)==union.combo(a,seq(2)+1)$i))
stopifnot(all(seq(4)+2==union.combo(b,6)$i))
stopifnot(all(seq(6)==union.combo(a,b,6)$i))
}
