% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{urandCap}
\alias{urandCap}
\title{urandCap}
\usage{
urandCap(cap, n = 1, ...)
}
\arguments{
\item{cap}{a 'numeric' or 'bigz' positive integer cap}

\item{n}{a 'numeric' positive integer number of observations to produce}
}
\description{
Generate random integers in the range 1:cap
}
